Workaround for Pillow red artifacts
Some preliminary tests with the US National Map WMS service were showing
red artifacts like these ones:

https://pbs.twimg.com/media/FDxTWJOXoAQ3aTq?format=jpg&name=large

Upon further investigation it turned out that Pillow was adding red
pixels to the mask image created from the numpy array. For instance for
this command:

        munibot -i 53029 create us

In `process_image()` we get:

        mask_array[1200][1019]

        [(0, 0, 0), dtype='uint8']

But after creating the `mask` image with `mask = Image.fromarray(mask_array)`,
we get:

        pixels = mask.load()

        pixels[1200, 1019]

        (255, 0, 0)

I have no idea why this happens, but it looks like a bug in Pillow. I
modified the code to ensure that all RGB pixels are equal when setting
the alpha value and that should take care of the issue.